@@ -1,56 +1,44 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
-  <PropertyGroup>
-    <TargetFramework>netcoreapp1.1</TargetFramework>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>netcoreapp3.1</TargetFramework>
+	</PropertyGroup>
 
-  <PropertyGroup>
-    <PackageTargetFallback>$(PackageTargetFallback);portable-net45+win8+wp8+wpa81;</PackageTargetFallback>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <UserSecretsId>aspnet-Wilson.Web-21879d00-3554-4155-ae27-e65eb872770f</UserSecretsId>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Content Remove="wwwroot/css\navbar-fixed-side.scss" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="AutoMapper" Version="6.0.2" />
-    <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.0.0" />
-    <PackageReference Include="Microsoft.AspNetCore" Version="1.1.2" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.Cookies" Version="1.1.2" />
-    <PackageReference Include="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" Version="1.1.2" />
-    <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="1.1.2" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="1.1.3" />
-    <PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="1.1.2" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="1.1.2" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="1.1.2" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer.Design" Version="1.1.2" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="1.1.1" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="1.1.2" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="1.1.2" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="1.1.1" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.BrowserLink" Version="1.1.2" />
-  </ItemGroup>
-  <ItemGroup>
-    <DotNetCliToolReference Include="Microsoft.EntityFrameworkCore.Tools.DotNet" Version="1.0.0" />
-    <DotNetCliToolReference Include="Microsoft.Extensions.SecretManager.Tools" Version="1.0.0" />
-    <DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="1.0.0" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Areas\Companies\Views\Shared\DisplayTemplates\" />
-    <Folder Include="Views\Install\" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\Accounting\Wilson.Accounting.Core\Wilson.Accounting.Core.csproj" />
-    <ProjectReference Include="..\..\Accounting\Wilson.Accounting.Data\Wilson.Accounting.Data.csproj" />
-    <ProjectReference Include="..\..\Companies\Wilson.Companies.Core\Wilson.Companies.Core.csproj" />
-    <ProjectReference Include="..\..\Companies\Wilson.Companies.Data\Wilson.Companies.Data.csproj" />
-    <ProjectReference Include="..\..\Projects\Wilson.Projects.Core\Wilson.Projects.Core.csproj" />
-    <ProjectReference Include="..\..\Projects\Wilson.Projects.Data\Wilson.Projects.Data.csproj" />
-    <ProjectReference Include="..\..\Scheduler\Wilson.Scheduler.Core\Wilson.Scheduler.Core.csproj" />
-    <ProjectReference Include="..\..\Scheduler\Wilson.Scheduler.Data\Wilson.Scheduler.Data.csproj" />
-  </ItemGroup>
-
+	<ItemGroup>
+		<PackageReference Include="AutoMapper" Version="6.2.2" />
+		<PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.15.0" />
+		<PackageReference Include="Microsoft.AspNetCore" Version="2.2.0" />
+		<PackageReference Include="Microsoft.AspNetCore.Authentication.Cookies" Version="2.2.0" />
+		<PackageReference Include="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" Version="3.1.9" />
+		<PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="3.1.9" />
+		<PackageReference Include="Microsoft.AspNetCore.Mvc" Version="2.2.0" />
+		<PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="2.2.0" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="3.1.9" PrivateAssets="All" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="3.1.9" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer.Design" Version="1.1.6" PrivateAssets="All" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="3.1.9" PrivateAssets="All" />
+		<PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="3.1.9" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="3.1.9" />
+		<PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="3.1.4" PrivateAssets="All" />
+		<PackageReference Include="Microsoft.VisualStudio.Web.BrowserLink" Version="2.2.0" />
+	</ItemGroup>
+	<ItemGroup>
+		<DotNetCliToolReference Include="Microsoft.EntityFrameworkCore.Tools.DotNet" Version="1.0.0" />
+		<DotNetCliToolReference Include="Microsoft.Extensions.SecretManager.Tools" Version="1.0.0" />
+		<DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="1.0.0" />
+	</ItemGroup>
+	<ItemGroup>
+		<Folder Include="Areas\Companies\Views\Shared\DisplayTemplates\" />
+		<Folder Include="Views\Install\" />
+	</ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\..\Accounting\Wilson.Accounting.Core\Wilson.Accounting.Core.csproj" />
+		<ProjectReference Include="..\..\Accounting\Wilson.Accounting.Data\Wilson.Accounting.Data.csproj" />
+		<ProjectReference Include="..\..\Companies\Wilson.Companies.Core\Wilson.Companies.Core.csproj" />
+		<ProjectReference Include="..\..\Companies\Wilson.Companies.Data\Wilson.Companies.Data.csproj" />
+		<ProjectReference Include="..\..\Projects\Wilson.Projects.Core\Wilson.Projects.Core.csproj" />
+		<ProjectReference Include="..\..\Projects\Wilson.Projects.Data\Wilson.Projects.Data.csproj" />
+		<ProjectReference Include="..\..\Scheduler\Wilson.Scheduler.Core\Wilson.Scheduler.Core.csproj" />
+		<ProjectReference Include="..\..\Scheduler\Wilson.Scheduler.Data\Wilson.Scheduler.Data.csproj" />
+	</ItemGroup>
 </Project>
