@@ -1,4 +1,8 @@
-﻿namespace Wilson.Web.Areas.Companies.Models.InquiriesViewModels
+﻿using System;
+using System.Collections.Generic;
+using Wilson.Companies.Core.Enumerations;
+
+namespace Wilson.Web.Areas.Companies.Models.InquiriesViewModels
 {
     public class EmployeeViewModel
     {
@@ -11,6 +15,16 @@
         public override string ToString()
         {
             return this.FirstName + " " + this.LastName;
+        }
+
+        public string Email { get; set; }
+        public string Phone { get; set; }
+        public string PrivatePhone { get; set; }
+        public string EmployeePosition { get; set; }
+
+        public Array EmployeePositions()
+        {
+            return Enum.GetValues(typeof(Wilson.Companies.Core.Enumerations.EmployeePosition));
         }
     }
 }
