@@ -1,6 +1,5 @@
 ﻿@using System.Collections.Generic
 @using Microsoft.AspNetCore.Http
-@using Microsoft.AspNetCore.Http.Authentication
 @using Wilson.Companies.Core.Entities
 
 @model LoginViewModel
@@ -13,10 +12,10 @@
 <h2>WILSON</h2>
 <h4>Enterprise Resource Planing System</h4>
 <hr />
-<div class="row">    
-    <div class="col-md-8 col-lg-offset-3">        
+<div class="row">
+    <div class="col-md-8 col-lg-offset-3">
         <section>
-            <form asp-controller="Account" asp-action="Login" asp-route-returnurl="@ViewData["ReturnUrl"]" method="post" class="form-horizontal">                
+            <form asp-controller="Account" asp-action="Login" asp-route-returnurl="@ViewData["ReturnUrl"]" method="post" class="form-horizontal">
                 <div asp-validation-summary="All" class="text-danger"></div>
                 <div class="form-group">
                     <label asp-for="Email" class="col-md-2 control-label"></label>
@@ -46,7 +45,7 @@
                     <div class="col-md-offset-2 col-md-10">
                         <button type="submit" class="btn btn-default">Log in</button>
                     </div>
-                </div>                
+                </div>
             </form>
         </section>
     </div>
